Silence clang-tidy `signed char` to `int` conversion warning
> warning: 'signed char' to 'int' conversion; consider casting to
> 'unsigned char' first. [bugprone-signed-char-misuse,cert-str34-c]

As can be seen in the clang-tidy documentation, this warning is
completely irrelevant to the code in question:

https://releases.llvm.org/17.0.1/tools/clang/tools/extra/docs/clang-tidy/checks/bugprone/signed-char-misuse.html

+++ src/util/utf8.c
@@ -31,7 +31,7 @@ static const int8_t seq_len_table[256] = {
 
 static int u_seq_len(unsigned char first_byte)
 {
-    int len = seq_len_table[first_byte];
+    int8_t len = seq_len_table[first_byte];
     BUG_ON(len < I || len > UTF8_MAX_SEQ_LEN);
     return len;
 }
